rename 'syslog1' to 'syslog_tst'
